Рабочая версия игры

+++ Form1.cs
@@ -13,12 +13,19 @@ namespace Belive
         private bool _isGameMode;
         private string _path = "data.txt";
         private bool _isCorrectValue = true;
-        private int _trueValue = 0;
-        private int _falseValue = 0;
+        private int _trueValue;
+        private int _falseValue;
+        private int _valueCountDb = 0;
 
         public MainForm()
         {
             InitializeComponent();
+            Qlbl.Parent = ImageBoxQ;
+            Qlbl.Dock = DockStyle.Fill;
+            Albl.Parent = ImageBoxA;
+            Albl.Dock = DockStyle.Fill;
+            GamePanel.Visible = false;
+            EditorPanel.Visible = false;
         }
 
         private void Exit_Click(object sender, EventArgs e)
@@ -58,23 +65,18 @@ namespace Belive
 
         private void NumQuestion_ValueChanged(object sender, EventArgs e)
         {
-            if (_db != null)
-            {
-                QuestionLbl.Text = _db[(int)NumQuestion.Value - 1].TextQuestion;
-                TrueFalseCheck.Checked = _db[(int)NumQuestion.Value - 1].IsTrue;
-                ImageBox.ImageLocation = _db[(int)NumQuestion.Value - 1].PathImageQuestion;
-                UIPanel.Parent = ImageBox;
-                _isCorrectValue = true;
-            }
+            if (IsDataBaseNoCreate()) return;
+            Qlbl.Text = _db[(int)NumQuestion.Value - 1].TextQuestion;
+            Albl.Text = _db[(int) NumQuestion.Value - 1].TextAnswer;
+            TrueFalseCheck.Checked = _db[(int)NumQuestion.Value - 1].IsTrue;
+            ImageBoxQ.ImageLocation = _db[(int) NumQuestion.Value - 1].PathImageQuestion;
+            ImageBoxA.ImageLocation = _db[(int)NumQuestion.Value - 1].PathImageAnswer;
+            _isCorrectValue = true;
         }
 
         private void AddBtn_Click(object sender, EventArgs e)
         {
-            if (_db == null)
-            {
-                MessageBox.Show(@"Создайте новую базу данных", @"Сообщение");
-                return;
-            }
+            if (IsDataBaseNoCreate()) return;
             _db.Add((_db.Count + 1).ToString(), string.Empty, true, string.Empty, string.Empty);
             NumQuestion.Maximum = _db.Count;
             NumQuestion.Value = _db.Count;
@@ -82,6 +84,8 @@ namespace Belive
 
         private void DelBtn_Click(object sender, EventArgs e)
         {
+            if (IsDataBaseNoCreate()) return;
+
             if (NumQuestion.Maximum == 1 || _db == null) return;
             _db.Remove((int)NumQuestion.Value);
             NumQuestion.Maximum--;
@@ -97,12 +101,6 @@ namespace Belive
 
         private void OpenFile_Click(object sender, EventArgs e)
         {
-            LoadPathDb(_path);
-            if (_pathSave != string.Empty && _pathSave != "")
-            {
-                LoadParam(_pathSave);
-                return;
-            }
             OpenFileDialog ofd = new OpenFileDialog();
             if (ofd.ShowDialog() == DialogResult.OK)
             {
@@ -112,32 +110,50 @@ namespace Belive
 
         private void LoadParam(string path)
         {
-            _db = new TrueFalse(path);
-            _db.Load();
-            NumQuestion.Minimum = 1;
-            NumQuestion.Maximum = _db.Count;
-            NumQuestion.Value = 1;
-            QuestionLbl.Parent = ImageBox;
-            QuestionLbl.BackColor = Color.Transparent;
-            TrueCoutLbl.Text = _trueValue.ToString();
-            TrueCoutLbl.Text = _falseValue.ToString();
+            try
+            {
+                _db = new TrueFalse(path);
+                _db.Load();
+                NumQuestion.Minimum = 1;
+                NumQuestion.Maximum = _db.Count;
+                NumQuestion.Value = 1;
+            }
+            catch
+            {
+                MessageBox.Show("Базаданных не загружена");
+            }
         }
 
         private void SaveBtn_Click(object sender, EventArgs e)
         {
-            if (_db == null) return;
+            if (IsDataBaseNoCreate()) return;
             _db[(int)NumQuestion.Value - 1].TextQuestion = QuestionTextBox.Text;
             _db[(int)NumQuestion.Value - 1].IsTrue = TrueFalseCheck.Checked;
+            Qlbl.Text = QuestionTextBox.Text;
         }
 
         private void AddImgBtn_Click(object sender, EventArgs e)
         {
+            if (IsDataBaseNoCreate()) return;
+
             OpenFileDialog ofd = new OpenFileDialog();
             if (ofd.ShowDialog() == DialogResult.OK)
             {
-                ImageBox.ImageLocation = ofd.FileName;
-                _db[(int)NumQuestion.Value - 1].PathImageQuestion = ImageBox.ImageLocation;
+                var path = ImageBoxQ.ImageLocation = ofd.FileName;
+                ImageBoxQ.Image = Image.FromFile(path);
+                _db[(int)NumQuestion.Value - 1].PathImageQuestion = path;
             }
+        }
+
+        private bool IsDataBaseNoCreate()
+        {
+            if (_db == null)
+            {
+                MessageBox.Show(@"Для корректной работы необходимо открыть сохраненную базу данных или создать новую.");
+                return true;
+            }
+
+            return false;
         }
 
         private void SwithMode_Click(object sender, EventArgs e)
@@ -149,34 +165,91 @@ namespace Belive
 
         private void ImgAnsBtn_Click(object sender, EventArgs e)
         {
+            if (IsDataBaseNoCreate()) return;
+
             OpenFileDialog ofd = new OpenFileDialog();
             if (ofd.ShowDialog() == DialogResult.OK)
             {
-                ImageBox.ImageLocation = ofd.FileName;
-                _db[(int)NumQuestion.Value - 1].PathImageAnswer = ImageBox.ImageLocation;
-            }
-        }
-
-        private void TrueBtn_Click(object sender, EventArgs e)
-        {
-            if (_db != null && _isCorrectValue)
-            {
-                ImageBox.ImageLocation = _db[(int) NumQuestion.Value - 1].PathImageAnswer;
-                QuestionLbl.Text = _db[(int) NumQuestion.Value - 1].TextAnswer;
-                if (_db[(int) NumQuestion.Value - 1].IsTrue)
-                    _trueValue++;
-                else
-                    _falseValue++;
-                TrueCoutLbl.Text = $@"{_trueValue}";
-                FalseCountLbl.Text = $@"{_falseValue}";
-                _isCorrectValue = !_isCorrectValue;
+                ImageBoxA.ImageLocation = ofd.FileName;
+                _db[(int)NumQuestion.Value - 1].PathImageAnswer = ImageBoxA.ImageLocation;
             }
         }
 
         private void SaveBtnA_Click(object sender, EventArgs e)
         {
-            if (_db == null) return;
+            if (IsDataBaseNoCreate()) return;
             _db[(int)NumQuestion.Value - 1].TextAnswer = QuestionTextBox.Text;
+            _db[(int)NumQuestion.Value - 1].IsTrue = TrueFalseCheck.Checked;
+            Albl.Text = QuestionTextBox.Text;
+        }
+
+        private void PlayBtn_Click(object sender, EventArgs e)
+        {
+            GamePanel.Visible = true;
+            MainMenuPanel.Visible = false;
+            EditorPanel.Visible = false;
+        }
+
+        private void ExitMainMenuBtn_Click(object sender, EventArgs e)
+        {
+            GamePanel.Visible = false;
+            MainMenuPanel.Visible = true;
+        }
+
+        private void MoveEditorBtn_Click(object sender, EventArgs e)
+        {
+            GamePanel.Visible = false;
+            MainMenuPanel.Visible = false;
+            EditorPanel.Visible = true;
+        }
+
+        private void ReturnMenuBtn_Click(object sender, EventArgs e)
+        {
+            GamePanel.Visible = false;
+            MainMenuPanel.Visible = true;
+            EditorPanel.Visible = false;
+        }
+
+        private void ExitBtnMain_Click(object sender, EventArgs e)
+        {
+            Close();
+        }
+
+        private void ContinueBtn_Click(object sender, EventArgs e)
+        {
+            LoadPathDb(_path);
+            if (_pathSave != string.Empty && _pathSave != "")
+            {
+                LoadParam(_pathSave);
+            }
+            TrueBtn.Visible = true;
+            FalseBtn.Visible = true;
+            ContinueBtn.Visible = false;
+            GamePictureBox.ImageLocation = _db[_valueCountDb].PathImageQuestion;
+        }
+
+        private void TrueBtn_Click(object sender, EventArgs e)
+        {
+            GamePictureBox.ImageLocation = _db[_valueCountDb].PathImageAnswer;
+            if (_valueCountDb < _db.Count - 1)
+                _valueCountDb++;
+            TrueBtn.Visible = false;
+            FalseBtn.Visible = false;
+            ContinueBtn.Visible = true;
+        }
+
+        private void FalseBtn_Click(object sender, EventArgs e)
+        {
+            GamePictureBox.ImageLocation = _db[_valueCountDb].PathImageAnswer;
+            if (_valueCountDb < _db.Count - 1)
+                _valueCountDb++;
+            else
+            {
+                MessageBox.Show(@"Игра окончена");
+            }
+            TrueBtn.Visible = false;
+            FalseBtn.Visible = false;
+            ContinueBtn.Visible = true;
         }
     }
 }
